updating ecommerce

+++ week6/ecommerce/src/main/java/com/tts/ecommerce/controller/ProductController.java
@@ -1,9 +1,10 @@
 package com.tts.ecommerce.controller;
 
+import com.tts.ecommerce.model.Product;
+import com.tts.ecommerce.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -11,6 +12,8 @@ import org.springframework.web.bind.annotation.RequestMethod;
 
 import javax.validation.Valid;
 
+// this is our controller for our product page
+// note that we have a method that allows us to create or update a product
 @Controller
 public class ProductController {
 
@@ -18,14 +21,14 @@ public class ProductController {
     ProductService productService;
 
     @GetMapping("/product/{id}")
-    public String show(@PathVariable int id, Model model){
+    public String show(@PathVariable long id, Model model) {
         Product product = productService.findById(id);
         model.addAttribute(product);
         return "product";
     }
 
-    @RequestMapping(value="/product", method = {RequestMethod.POST,RequestMethod.PUT})
-    public String createOrUpdate(@Valid Product product){
+    @RequestMapping(value = "/product", method = {RequestMethod.POST, RequestMethod.PUT})
+    public String createOrUpdate(@Valid Product product) {
         productService.save(product);
         return "redirect:/product/" + product.getId();
     }
